philo

+++ philo/philo.h
@@ -6,7 +6,7 @@
 /*   By: schahir <schahir@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/24 17:39:10 by schahir           #+#    #+#             */
-/*   Updated: 2025/08/09 21:30:03 by schahir          ###   ########.fr       */
+/*   Updated: 2025/08/10 11:22:10 by schahir          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -55,7 +55,7 @@ void	*routine(void *data);
 int	print_routine(t_philo *philo, char *msg);
 int  check_n_delay(t_philo *philo);
 /*-------Utils----------*/
-long				ft_atoi(char *str);
+int				ft_atoi(char *str);
 void				*ft_calloc(size_t nmemb, size_t size);
 void				putstr_fd(char *s, int fd);
 void				print_n_destroy(t_schedule *s, char *msg);
